Redefine TraceRegistry as a concrete class as the interface currently had only one implementation.
