+++ profiler/src/test/java/com/splunk/opentelemetry/profiler/snapshot/TraceRegistrationTest.java
@@ -22,7 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class TraceRegistrationTest {
-  private final DefaultTraceRegistry registry = new DefaultTraceRegistry();
+  private final TraceRegistry registry = new TraceRegistry();
   private final SnapshotProfilingSdkCustomizer customizer = new SnapshotProfilingSdkCustomizer(registry);
 
   @RegisterExtension
